+++ cmd/gctsExecuteABAPQualityChecks_generated_test.go
@@ -1,0 +1,17 @@
+package cmd
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestGctsExecuteABAPQualityChecksCommand(t *testing.T) {
+	t.Parallel()
+
+	testCmd := GctsExecuteABAPQualityChecksCommand()
+
+	// only high level testing performed - details are tested in step generation procedure
+	assert.Equal(t, "gctsExecuteABAPQualityChecks", testCmd.Use, "command name incorrect")
+
+}
