+++ documentation/mkdocs.yml
@@ -95,7 +95,7 @@ nav:
         - gctsCloneRepository: steps/gctsCloneRepository.md
         - gctsCreateRepository: steps/gctsCreateRepository.md
         - gctsDeploy: steps/gctsDeploy.md
-        - gctsExecuteABAPUnitTests: steps/gctsExecuteABAPUnitTests.md
+        - gctsExecuteABAPQualityChecks: steps/gctsExecuteABAPQualityChecks.md
         - gctsRollback: steps/gctsRollback.md
         - githubCheckBranchProtection: steps/githubCheckBranchProtection.md
         - githubCommentIssue: steps/githubCommentIssue.md
@@ -171,6 +171,7 @@ nav:
         - checkChangeInDevelopment: steps/checkChangeInDevelopment.md
         - npmExecute: steps/npmExecute.md
         - transportRequestUploadFile: steps/transportRequestUploadFile.md
+        - gctsExecuteABAPUnitTests: steps/gctsExecuteABAPUnitTests.md
     - 'Command line tool': cli/index.md
 
 theme:
