+++ vars/gctsExecuteABAPQualityChecks.groovy
@@ -1,0 +1,11 @@
+import groovy.transform.Field
+
+@Field String STEP_NAME = getClass().getName()
+@Field String METADATA_FILE = 'metadata/gctsExecuteABAPQualityChecks.yaml'
+
+void call(Map parameters = [:]) {
+        List credentials = [
+        [type: 'usernamePassword', id: 'abapCredentialsId', env: ['PIPER_username', 'PIPER_password']]
+        ]
+        piperExecuteBin(parameters, STEP_NAME, METADATA_FILE, credentials)
+}
